ra_ultimate.py: fix BF16 dtype mismatch in benchmark
The benchmark function was creating input tensor in FP32 while
models were converted to BF16, causing dtype mismatch errors
during autocast. Fixed by explicitly creating input tensor with
dtype=torch.bfloat16 to match model precision.

This completes the ultimate RA implementation combining all
optimizations: same-FLOP approach (D_std=56, R=8), zero-copy
folding, fused bias, symmetric Q/K scaling, head-selective
routing, BF16 precision, and forced Flash Attention backend.

Expected performance with head-selective routing: 2.0-2.1ms
(match baseline SDPA speed).

Also added comprehensive documentation:
- RA_ULTIMATE_ANALYSIS.md: technical deep-dive of optimization
  journey from 4.6x overhead to expected 1.0-1.35x
- NEXT_STEPS.md: actionable plan for GPU testing, quality
  validation, and integration

Generated-by: Claude AI

+++ ra_ultimate.py
@@ -232,7 +232,7 @@ def benchmark_ultimate():
     print("  - BF16 + Flash backend")
     print()
 
-    x = torch.randn(B, T, n_embd, device=device)
+    x = torch.randn(B, T, n_embd, device=device, dtype=torch.bfloat16)
 
     # Baseline for comparison
     print("1. Baseline SDPA (optimized)...")
